Move node types code to its own module
